cleanup test dir
