+++ packages/remix/test/integration/app_v2/routes/capture-exception.tsx
@@ -1,2 +1,7 @@
-export * from '../../common/routes/capture-exception';
-export { default } from '../../common/routes/capture-exception';
+import * as Sentry from '@sentry/remix';
+
+export default function ErrorBoundaryCapture() {
+  Sentry.captureException(new Error('Sentry Manually Captured Error'));
+
+  return <div></div>;
+}
