+++ packages/remix/test/integration/app_v2/routes/capture-message.tsx
@@ -1,2 +1,7 @@
-export * from '../../common/routes/capture-message';
-export { default } from '../../common/routes/capture-message';
+import * as Sentry from '@sentry/remix';
+
+export default function ErrorBoundaryCapture() {
+  Sentry.captureMessage('Sentry Manually Captured Message');
+
+  return <div />;
+}
