+++ packages/remix/test/integration/app_v2/routes/error-boundary-capture.$id.tsx
@@ -1,2 +1,13 @@
-export * from '../../common/routes/error-boundary-capture.$id';
-export { default } from '../../common/routes/error-boundary-capture.$id';
+import { useState } from 'react';
+
+export default function ErrorBoundaryCapture() {
+  const [count, setCount] = useState(0);
+
+  if (count > 0) {
+    throw new Error('Sentry React Component Error');
+  } else {
+    setTimeout(() => setCount(count + 1), 0);
+  }
+
+  return <div>{count}</div>;
+}
