+++ packages/remix/test/integration/app_v2/routes/index.tsx
@@ -1,2 +1,7 @@
-export * from '../../common/routes/index';
-export { default } from '../../common/routes/index';
+export default function Index() {
+  return (
+    <div>
+      <h1>Remix Integration Tests Home</h1>
+    </div>
+  );
+}
