+++ packages/remix/test/integration/app_v2/routes/loader-defer-response.$id.tsx
@@ -1,2 +1,18 @@
-export * from '../../common/routes/loader-defer-response.$id';
-export { default } from '../../common/routes/loader-defer-response.$id';
+import { LoaderFunctionArgs, defer } from '@remix-run/node';
+import { useLoaderData } from '@remix-run/react';
+
+export const loader = async ({ params: { id } }: LoaderFunctionArgs) => {
+  return defer({
+    id,
+  });
+};
+
+export default function LoaderJSONResponse() {
+  const data = useLoaderData<typeof loader>();
+
+  return (
+    <div>
+      <h1 id="data-render">{data?.id ? data.id : 'Not Found'}</h1>
+    </div>
+  );
+}
