+++ packages/remix/test/integration/app_v2/routes/manual-tracing.$id.tsx
@@ -1,2 +1,10 @@
-export * from '../../common/routes/manual-tracing.$id';
-export { default } from '../../common/routes/manual-tracing.$id';
+import * as Sentry from '@sentry/remix';
+
+export default function ManualTracing() {
+  const span = Sentry.startInactiveSpan({
+    name: 'test_transaction_1',
+    forceTransaction: true,
+  });
+  span.end();
+  return <div />;
+}
