+++ packages/remix/test/integration/app_v2/routes/scope-bleed.$id.tsx
@@ -1,2 +1,18 @@
-export * from '../../common/routes/scope-bleed.$id';
-export { default } from '../../common/routes/scope-bleed.$id';
+import { LoaderFunction, json } from '@remix-run/node';
+
+import * as Sentry from '@sentry/remix';
+
+export const loader: LoaderFunction = async ({ params: { id } }) => {
+  const timeTil = parseInt(id || '', 10) * 1000;
+  await new Promise(resolve => setTimeout(resolve, 3000 - timeTil));
+  Sentry.setTag(`tag${id}`, id);
+  return json({ test: 'test' });
+};
+
+export default function ScopeBleed() {
+  return (
+    <div>
+      <h1>Hello</h1>
+    </div>
+  );
+}
