+++ packages/remix/test/integration/app_v2/routes/throw-redirect.tsx
@@ -1,2 +1,11 @@
-export * from '../../common/routes/throw-redirect';
-export { default } from '../../common/routes/throw-redirect';
+import { LoaderFunction, redirect } from '@remix-run/node';
+import { useLoaderData } from '@remix-run/react';
+
+export const loader: LoaderFunction = async () => {
+  throw redirect('/');
+};
+
+export default function ThrowRedirect() {
+  const data = useLoaderData();
+  return <div>{data}</div>;
+}
