--- packages/remix/test/integration/common/routes/capture-exception.tsx
@@ -1,7 +1,0 @@
-import * as Sentry from '@sentry/remix';
-
-export default function ErrorBoundaryCapture() {
-  Sentry.captureException(new Error('Sentry Manually Captured Error'));
-
-  return <div></div>;
-}
