--- packages/remix/test/integration/common/routes/capture-message.tsx
@@ -1,7 +1,0 @@
-import * as Sentry from '@sentry/remix';
-
-export default function ErrorBoundaryCapture() {
-  Sentry.captureMessage('Sentry Manually Captured Message');
-
-  return <div />;
-}
