--- packages/remix/test/integration/common/routes/index.tsx
@@ -1,7 +1,0 @@
-export default function Index() {
-  return (
-    <div>
-      <h1>Remix Integration Tests Home</h1>
-    </div>
-  );
-}
