--- packages/remix/test/integration/common/routes/ssr-error.tsx
@@ -1,7 +1,0 @@
-export default function SSRError() {
-  const data = ['err'].map(err => {
-    throw new Error('Sentry SSR Test Error');
-  });
-
-  return <div>{data}</div>;
-}
