--- packages/remix/test/integration/common/routes/throw-redirect.tsx
@@ -1,11 +1,0 @@
-import { LoaderFunction, redirect } from '@remix-run/node';
-import { useLoaderData } from '@remix-run/react';
-
-export const loader: LoaderFunction = async () => {
-  throw redirect('/');
-};
-
-export default function ThrowRedirect() {
-  const data = useLoaderData();
-  return <div>{data}</div>;
-}
